+++ Source/Server/Game/Objects/Item.cs
@@ -205,14 +205,14 @@ public static class Item
             return;
         }
 
-        var user = IsEditorLocked(session.Id, (byte) EditorType.Item);
+        var user = IsEditorLocked(session.Id, EditorType.Item);
         if (!string.IsNullOrEmpty(user))
         {
             NetworkSend.PlayerMsg(session.Id, "The game editor is locked and being used by " + user + ".", (int) Color.BrightRed);
             return;
         }
 
-        Data.TempPlayer[session.Id].Editor = (byte) EditorType.Item;
+        Data.TempPlayer[session.Id].Editor = EditorType.Item;
 
         Animation.SendAnimations(session.Id);
         Projectile.SendProjectiles(session.Id);
