+++ Source/Server/Game/Objects/Npc.cs
@@ -499,14 +499,14 @@ public static class Npc
             return;
         }
 
-        var user = IsEditorLocked(session.Id, (byte) EditorType.Npc);
+        var user = IsEditorLocked(session.Id, EditorType.Npc);
         if (!string.IsNullOrEmpty(user))
         {
             NetworkSend.PlayerMsg(session.Id, "The game editor is locked and being used by " + user + ".", (int) Color.BrightRed);
             return;
         }
 
-        Data.TempPlayer[session.Id].Editor = (byte) EditorType.Npc;
+        Data.TempPlayer[session.Id].Editor = EditorType.Npc;
 
         Item.SendItems(session.Id);
         Animation.SendAnimations(session.Id);
