+++ Source/Server/Game/Objects/Resource.cs
@@ -120,14 +120,14 @@ public static class Resource
             return;
         }
 
-        var user = IsEditorLocked(session.Id, (byte) EditorType.Resource);
+        var user = IsEditorLocked(session.Id, EditorType.Resource);
         if (!string.IsNullOrEmpty(user))
         {
             NetworkSend.PlayerMsg(session.Id, "The game editor is locked and being used by " + user + ".", (int) Color.BrightRed);
             return;
         }
 
-        Data.TempPlayer[session.Id].Editor = (byte) EditorType.Resource;
+        Data.TempPlayer[session.Id].Editor = EditorType.Resource;
 
         Item.SendItems(session.Id);
         Animation.SendAnimations(session.Id);
