+++ Source/Server/Game/Objects/Script.cs
@@ -23,14 +23,14 @@ public static class Script
             return;
         }
 
-        var user = IsEditorLocked(session.Id, (byte) EditorType.Script);
+        var user = IsEditorLocked(session.Id, EditorType.Script);
         if (!string.IsNullOrEmpty(user))
         {
             NetworkSend.PlayerMsg(session.Id, "The game editor is locked and being used by " + user + ".", (int) Color.BrightRed);
             return;
         }
 
-        Data.TempPlayer[session.Id].Editor = (byte) EditorType.Script;
+        Data.TempPlayer[session.Id].Editor = EditorType.Script;
 
         var codeLines = Data.Script.Code ?? [];
